Update function file rename
