+++ functions/create_draft/interactivity_handler.ts
@@ -3,7 +3,7 @@ import {
   ViewSubmissionHandler,
 } from "deno-slack-sdk/functions/interactivity/types.ts";
 
-import { CreateDraftFunctionDefinition as CreateDraftFunction } from "./mod.ts";
+import { CreateDraftFunctionDefinition as CreateDraftFunction } from "./definition.ts";
 import {
   buildConfirmSendModal,
   buildDraftBlocks,
