+++ functions/send_announcement/definition.ts
